updated TableIn to be violation invariants
managed to reduce outputs to a single violation var
